Set remote date to current local time
Usually, I'm working with a Pi that doesn't have an internet connection.
That means its clock is most likely
wrong - so I should fix that
before attempting anything that
might, for example, need to verify
SSL certificates.
This adds a quick and dirty command to set the date
before it starts a shell.